pyproject.toml
[project]
name = "spatial-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fpdf2>=2.8.5",
    "google-adk>=1.22.0",
    "google-genai>=1.57.0",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
]
